Unify logframe components
