+++ components/logframe/outputs-card.tsx
