+++ components/logframe/server-actions.ts
@@ -1,7 +1,26 @@
 'use server';
 
 import { createClient } from '@/utils/supabase/server';
-import { OutputMeasurable } from '@/utils/types';
+import {
+  Impact,
+  Outcome,
+  OutcomeMeasurable,
+  Output,
+  OutputMeasurable,
+} from '@/utils/types';
+
+export const fetchLogframe = async (identifier: number | string) => {
+  const supabase = await createClient();
+  const response = await supabase
+    .from('projects')
+    .select(
+      'id, slug, name, impacts(*), outcomes(*, outcome_measurables(*, outputs(*))),  outputs(*)',
+    )
+    .eq(typeof identifier === 'number' ? 'id' : 'slug', identifier)
+    .single();
+
+  return response;
+};
 
 export const fetchOutputByCode = async (code: string, projectSlug: string) => {
   const supabase = await createClient();
@@ -25,6 +44,117 @@ export const fetchOutputByCode = async (code: string, projectSlug: string) => {
 export const fetchImpactIndicators = async (projectId: number) => {
   const supabase = await createClient();
   const { data, error } = await supabase.from('impact_indicators').select('*');
+
+  if (error) throw error;
+  return data;
+};
+
+export const upsertImpact = async (impact: Partial<Impact>) => {
+  const supabase = await createClient();
+  const { data, error } = await supabase
+    .from('impacts')
+    .upsert(impact)
+    .select()
+    .single();
+
+  if (error) throw error;
+  return data;
+};
+
+export const upsertOutcome = async (outcome: Partial<Outcome>) => {
+  const supabase = await createClient();
+  const { data, error } = await supabase
+    .from('outcomes')
+    .upsert({
+      id: outcome.id,
+      project_id: outcome.project_id,
+      description: outcome.description,
+      code: outcome.code,
+    })
+    .select()
+    .single();
+
+  if (error) throw error;
+
+  // Handle outcome_measurables
+  if (outcome.outcome_measurables) {
+    const measurablesToUpsert = outcome.outcome_measurables.map(
+      (measurable) => ({
+        ...measurable,
+        outcome_id: data.id,
+      }),
+    );
+
+    const { error: measurablesError } = await supabase
+      .from('outcome_measurables')
+      .upsert(measurablesToUpsert);
+
+    if (measurablesError) throw measurablesError;
+  }
+
+  return data;
+};
+
+export const upsertOutcomeMeasurable = async (
+  measurable: Partial<OutcomeMeasurable>,
+) => {
+  const supabase = await createClient();
+
+  // Ensure we have required fields
+  if (!measurable.outcome_id || !measurable.description || !measurable.code) {
+    throw new Error('Missing required fields');
+  }
+
+  const { data, error } = await supabase
+    .from('outcome_measurables')
+    .upsert({
+      id: measurable.id,
+      outcome_id: measurable.outcome_id,
+      description: measurable.description,
+      verification: measurable.verification || '',
+      assumptions: measurable.assumptions || '',
+      code: measurable.code,
+    })
+    .select()
+    .single();
+
+  if (error) throw error;
+  return data;
+};
+
+export const upsertOutput = async (output: Partial<Output>) => {
+  const supabase = await createClient();
+
+  if (!output.outcome_measurable_id || !output.description || !output.code) {
+    throw new Error('Missing required fields');
+  }
+
+  const { data: existingOutputs } = await supabase
+    .from('outputs')
+    .select('id, code')
+    .eq('project_id', output.project_id)
+    .neq('id', output.id || 0);
+
+  const isDuplicate = existingOutputs?.some(
+    (existingOutput) => existingOutput.code === output.code,
+  );
+
+  if (isDuplicate) {
+    throw new Error(`Output code "${output.code}" already exists`);
+  }
+
+  const { data, error } = await supabase
+    .from('outputs')
+    .upsert({
+      id: output.id,
+      outcome_measurable_id: output.outcome_measurable_id,
+      project_id: output.project_id,
+      description: output.description,
+      code: output.code,
+      status: output.status,
+    })
+    .select()
+    .single();
 
   if (error) throw error;
   return data;
